move 'map' tests to tests/map.ts
